requirements.txt
## Python Packages
pip
setuptools
raven
pytz
mysqlclient
coverage
selenium
#whitenoise
#django-haystack
#invoke
#pip-tools
#cookiecutter
#pytest
#sh
#elasticsearch
#logstash
#requests
#ipython
#ipdb
#flake8
#tablib
#WebTest
## Django Packages
django>=2.0.0,<2.1.0
django-tables2
django-filter
django-digg-paginator
django-debug-toolbar
django-configurations
#django-model-utils
#django-vanilla-views
#django-extensions
#django-widget-tweaks
#django-webtest
#model-mommy
#django-redis
#django-clear-cache
#django-mptt
#django-taggit
#elasticsearch_dsl
#gunicorn